--- src/main/java/com/example/labo5carry/Models/Commands/ICommand.java
@@ -1,9 +1,0 @@
-package com.example.labo5carry.Models.Commands;
-
-import com.example.labo5carry.Models.State;
-
-public interface ICommand {
-
-    void execute();
-    void setState(State state);
-}
