--- src/main/java/com/example/labo5carry/Models/Image.java
@@ -1,4 +1,0 @@
-package com.example.labo5carry.Models;
-
-public class Image {
-}
