+++ src/main/java/com/example/log121Lab05/Application.java
@@ -1,4 +1,4 @@
-package com.example.labo5carry;
+package com.example.log121Lab05;
 
 import javafx.fxml.FXMLLoader;
 import javafx.scene.Scene;
