+++ src/main/java/com/example/log121Lab05/IObserver.java
@@ -1,4 +1,4 @@
-package com.example.labo5carry;
+package com.example.log121Lab05;
 
 public interface IObserver
 {
