+++ src/main/java/com/example/log121Lab05/Models/Commands/Copy.java
@@ -1,8 +1,8 @@
-package com.example.labo5carry.Models.Commands;
+package com.example.log121Lab05.Models.Commands;
 
-import com.example.labo5carry.Controller;
-import com.example.labo5carry.Models.Perspective;
-import com.example.labo5carry.Models.State;
+import com.example.log121Lab05.Controller;
+import com.example.log121Lab05.Models.Perspective;
+import com.example.log121Lab05.Models.State;
 
 public class Copy implements ICommand {
