+++ src/main/java/com/example/log121Lab05/Models/Commands/ICommand.java
@@ -1,0 +1,9 @@
+package com.example.log121Lab05.Models.Commands;
+
+import com.example.log121Lab05.Models.State;
+
+public interface ICommand {
+
+    void execute();
+    void setState(State state);
+}
