+++ src/main/java/com/example/log121Lab05/Models/Commands/LoadImage.java
@@ -1,7 +1,7 @@
-package com.example.labo5carry.Models.Commands;
+package com.example.log121Lab05.Models.Commands;
 
-import com.example.labo5carry.Models.Services.FileService;
-import com.example.labo5carry.Models.State;
+import com.example.log121Lab05.Models.Services.FileService;
+import com.example.log121Lab05.Models.State;
 
 import javax.swing.*;
 import java.awt.image.BufferedImage;
