+++ src/main/java/com/example/log121Lab05/Models/Image.java
@@ -1,0 +1,4 @@
+package com.example.log121Lab05.Models;
+
+public class Image {
+}
